requirements.txt
streamlit
opencv-contrib-python-headless
pillow
numpy
